pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "georisk"
version = "0.1.0"
description = "Geopolitical early-warning forecasting platform"
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
  "lightgbm>=4.6.0",
  "numpy>=2.2.0",
  "pandas>=2.3.0",
  "pyarrow>=22.0.0",
  "pydantic>=2.12.0",
  "pyyaml>=6.0.0",
  "scikit-learn>=1.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.0",
]

[tool.setuptools]
packages = [
  "src",
  "src.common",
  "src.backtesting",
  "src.forecasting",
  "src.social_publishing",
  "src.website_publishing",
  "src.data_platform",
  "src.data_platform.ingestion",
  "src.data_platform.normalization",
  "src.data_platform.orchestration",
  "src.data_platform.serving",
  "src.data_platform.validation",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["src/tests"]
addopts = "-q"
